move compute_numerical_gradient method to separate class
